--- Songs.txt
@@ -1,175 +1,0 @@
-Chahun Main Ya Naa
-Sunn Raha Hai Na Tu
-Tum Hi Ho
-Hum Mar Jayenge
-Malare
-En Kadhal Solla
-No Tears Left To Cry
-Nenjukkul Peidhidum
-Oh! Shanthi Shanthi
-Chennai City Gangsta
-Oh Penne Penne
-Love Me Like You Do
-I Don't Wanna Live Forever
-Cheap Thrills
-Believer
-Shape of You
-3 Peg Sharry Mann
-Zingaat
-Sooraj Dooba Hain
-What Does The Fox Say
-A Thousand Years
-Hymn For The Weekend
-Closer
-Demons
-Radioactive
-Thunder
-On Top Of The World Imagine Dragons
-Bleeding Out
-Tiptoe Imagine Dragons
-Amsterdam
-Every Night Imagine Dragons
-Iridescent
-In The End
-Numb
-Castle of Glass
-One More Light
-Battle Symphony
-Heathens
-Sugar
-Daylight
-Payphone
-Animals Maroon 5
-It Ain't Me
-Attention
-How Long
-Stitches
-There's Nothing Holdin' Me Back
-Treat You Better
-Never Be Alone
-Counting Stars
-Passenger Let Her Go
-Thinking Out Loud
-Perfect Duet (With Beyonce)
-Castle On The Hill
-Love Story Taylor Swift
-Blank Space
-Gorgeous
-Baby Justin Bieber
-Let Me Love You
-Despacito Justin Bieber
-Sorry Justin Bieber
-Kar Gayi Chill
-Not Today Imagine Dragons
-Thuli Thuli
-Adra Adra Naakka Mukka
-New York Nagaram
-Ella Pugazhum
-Munbe Vaa
-Poongatre Poongatre
-Story Of My Life
-You & I
-Strong
-Something Just Like This 
-Don't Let Me Down
-Good Intentions Chainsmokers
-God Is A Woman
-Dangerous Woman
-Breathin
-This Love
-Havana
-Work From Home
-Thalli Pogathey
-Aaluma Doluma
-Kanave Kanave
-Iraiva Hiphop Tamizha
-Vaadi Pulla Vaadi
-Oxygen Song
-Enna Solla
-Dandanakka
-Barso Re
-Dilli Se Hu BC 
-Kala Chasma
-High Rated Gabru
-Priyaa na priyaa
-Lollypop Lageli
-Balam Pichkari
-Otha Sollala
-Happier Ed Sheeran Acoustic
-Girls Like You
-Whatever It Takes
-In My Blood
-Chittiyaan Kalaiyaan
-Symphony
-Perfect - One Direction
-Yamunai Aatrile
-End Game
-Thunderclouds
-Genius LSD
-Udhungada Sangu
-Score Unstoppable
-Elegy - Fat Rat
-Monody - Fat Rat
-Wolves
-Fight Song 
-Ovvoru Pookalume
-Mynaa Mynaa
-Mynaa Mynaa Nenjukkule
-Mere Rashqye Kamar
-Manali Trance
-Lungi Dance
-7 Years
-Lost In The Echo
-Kaiya Pudi
-Kabhi Kabhi Aditi
-Iragai Poley
-Dark Souls Forever Flame
-New Rules
-Bulleya Dil Hai Mushkil
-Hello - Adele
-Faded
-Catalyst Linkin Park
-She Doesn't Mind
-We Don't Talk Anymore
-Rockabye
-Centuries
-Human - RagnBone Man
-Dusk Till Dawn
-For You
-See You Again
-Uptown Girl - Billy Joel
-A Sky Full of Stars
-Love Yourself
-Do I Wanna Know
-Arabella
-Audio - LSD
-Konjum Mainaakale
-Pachai Nirame
-Nadhiye Nadhiye
-Yenna Solla Pogirai
-Chillena
-Oru Paadhi Kadhavu
-Sollitaley Ava Kaadhala
-Pookal Pookkum
-Vaamma Duraiyamma
-Just The Way You Are
-Badtameez Dil
-Let's Nacho
-The Humma Song
-Aise Na Mujhe Tum Dekho
-Vaseegara
-Photograph
-Let Me Love You - Ariana Grande
-Side To Side - Ariana Grande
-Diamond Heart - Alan Walker
-Dark Side - Alan Walker
-Zombie Bad Wolves
-Light em up - Fall Out Boys
-Play Me Like A Violin
-Crawling Linkin Park
-One Step Closer Linkin Park
-Heavy Linkin Park
-22 Taylor Swift
-Fight Song Shining
-Not Alone Alan Walker
-Sing Me To Sleep Alan Walker
