--- client_secret.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"777883687957-gde120f74mc1d8fb17uvaptn07vgohln.apps.googleusercontent.com","project_id":"song-snatcher","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://www.googleapis.com/oauth2/v3/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"3GbGEh7QVYS4ASRSO1Kgrcsr","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
