--- drive_credentials.json
@@ -1,1 +1,0 @@
-{"access_token": "ya29.GlstBpcXDjZCosJI0EqRJuJhZuB-3xz7eMzDiN1j7ZgS0iXJ8SUGFVLA6PNKT9VCPkaIKvUsboFP1E5YB6iyBKJkFhvSwURHzAc9Uv2Gr-RaKVQCDfW032DXhJJR", "client_id": "777883687957-gde120f74mc1d8fb17uvaptn07vgohln.apps.googleusercontent.com", "client_secret": "3GbGEh7QVYS4ASRSO1Kgrcsr", "refresh_token": "1/LyUJWOOHT_z06zhZjYs7gpy01ZIGryefCFKAVOkvbOQ", "token_expiry": "2018-10-05T12:48:52Z", "token_uri": "https://www.googleapis.com/oauth2/v3/token", "user_agent": null, "revoke_uri": "https://oauth2.googleapis.com/revoke", "id_token": null, "id_token_jwt": null, "token_response": {"access_token": "ya29.GlstBpcXDjZCosJI0EqRJuJhZuB-3xz7eMzDiN1j7ZgS0iXJ8SUGFVLA6PNKT9VCPkaIKvUsboFP1E5YB6iyBKJkFhvSwURHzAc9Uv2Gr-RaKVQCDfW032DXhJJR", "expires_in": 3600, "scope": "https://www.googleapis.com/auth/drive", "token_type": "Bearer"}, "scopes": ["https://www.googleapis.com/auth/drive"], "token_info_uri": "https://oauth2.googleapis.com/tokeninfo", "invalid": false, "_class": "OAuth2Credentials", "_module": "oauth2client.client"}
